+++ python/alpha93/packages/commons/src/alpha93/commons/__init__.py
@@ -1,9 +1,14 @@
 from ._scopes import catch, enumerate, throw
 from .types import typed
+from .utils import exceptions
 
 __all__ = (
     "catch",
     "enumerate",
+    "env",
+    "exceptions",
+    "futures",
+    "pydantic",
     "throw",
     "typed"
 )