+++ python/alpha93/packages/commons/src/alpha93/commons/pydantic/__init__.py
@@ -1,1 +1,5 @@
-from .types import Optional, Unset
+from .dataclasses import to_model
+
+__all__ = (
+    "to_model",
+)