+++ python/alpha93/packages/commons/src/alpha93/commons/utils/exceptions/_intrinsics.py
@@ -1,0 +1,7 @@
+def __intrinsics(identifier: str, /):
+    identifier = identifier + ": "
+    def __throw(message: str, /):
+        raise NotImplementedError(identifier + message)
+    return __throw
+
+TODO = __intrinsics("TODO")